requirements.txt
# ZMeta Stack requirements (runtime + dev/QA)
# --- Runtime ---
fastapi
uvicorn[standard]
pydantic>=2
pyyaml>=6
websockets>=12
requests>=2
tkintermapview>=1.29
python-dotenv>=1

# --- Dev/QA (optional) ---
httpx>=0.27
pytest>=7
black>=24
ruff>=0.5
mypy>=1.10
types-requests>=2
types-PyYAML>=6
